✨(tray) deploy multiple runners
We want to deploy multiple runners at once. For this, the deployment
contains multiple documents, one for each runner. a collection of
runners can be defined in the variable. The secrets should be used a
pattern where the runner name is used in its name. An example is
provided in the secret template.